Moved storybook files to story folder
