--- src/lib/components/Button/Button.stories.d.ts
@@ -1,5 +1,0 @@
-/// <reference types="react" />
-import { ComponentMeta } from '@storybook/react';
-declare const _default: ComponentMeta<({ children, disabled, primary, secondary, editor, outline, twClasses, onClick, type, }: import("./Button").ButtonProps) => JSX.Element>;
-export default _default;
-export declare const Primary: any;
