+++ src/stories/Button.stories.d.ts
@@ -1,0 +1,5 @@
+/// <reference types="react" />
+import { ComponentMeta } from '@storybook/react';
+declare const _default: ComponentMeta<({ children, disabled, primary, secondary, editor, outline, twClasses, onClick, type, }: import('../lib/components/Button/Button').ButtonProps) => JSX.Element>;
+export default _default;
+export declare const Primary: any;
