+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src/lib", "declaration.d.ts", "src/__test__/Button.spec.tsx"],
+  "include": ["src/lib", "declaration.d.ts", "src/__test__/Button.spec.tsx", "src/stories/Button.stories.tsx", "src/stories/Heading1.stories.tsx"],
   "exclude": ["node_modules", "dist"],
   "compilerOptions": {
     "allowJs": true,
